M4 Wave 6 AMS2: STRICT_COMPILERFLAGS for atkams/messages/lib, 1 genuine bug fixed
FossilOrigin-Name: b31d206d3200b932761251c7b3c14fccf35e3c35eaee2d42fd9ba37b8a4678b1

+++ src/atkams/messages/lib/amsn.c
@@ -381,7 +381,7 @@ long amsn__MS_DisambiguateFile(struct amsn *self, char *source, char *target, lo
     return(MS_DisambiguateFile(source, target, MustBeDir));
 }
 
-long amsn__MS_FastUpdateState(struct amsn *self)
+int amsn__MS_FastUpdateState(struct amsn *self)
 {
     return(MS_FastUpdateState());
 }
@@ -456,7 +456,7 @@ long amsn__MS_UnlinkFile(struct amsn *self, char *filename)
     return(MS_UnlinkFile(filename));
 }
 
-long amsn__MS_UpdateState(struct amsn *self)
+int amsn__MS_UpdateState(struct amsn *self)
 {
     return(MS_UpdateState());
 }
